sensor_client/requirements.txt
pyserial
pyzmq==26.0.3
typing
